src/app/hotels/hotels.service.ts
/* tslint:disable */
import { Injectable } from '@angular/core';
import { HttpClient } from '@angular/common/http';
import { environment } from '../../environments/environment';
import { Hotel } from './hotel.model';
import { Observable, throwError } from 'rxjs';
import { tap, catchError, map } from 'rxjs/operators';

@Injectable({
  providedIn: 'root'
})
export class HotelsService {
  baseUrl = environment.baseUrl;
  constructor(private http: HttpClient) { }

  getHotels(): Observable<Hotel[]> {
    return this.http.get<Hotel[]>(`${this.baseUrl}/tl0bp`).pipe(
      map(res => res['hotels']),
      catchError(err => throwError(err))
    );
  }
}
